+++ Back-end/FreteService/src/main/java/com/ecommerceproject/service/frete/service/FreteService.java
@@ -6,16 +6,26 @@ import org.springframework.stereotype.Service;
 
 /**
  * FreteService
+ * 
+ * Serviço que concentra as operaçoes e regras de negócio
+ * relacionadas ao frete.
  */
 @Service
 public class FreteService {
 
-    public int calcularValorFrete(int quatidadeItensSelecionados) {
+    /**
+     * Método que calcula o valor do frete a partir da quantidade
+     * de itens.
+     * 
+     * @param quatidadeItensSelecionados
+     * @return
+     */
+    public Double calcularValorFrete(int quatidadeItensSelecionados) {
         final int minValue = 5;
         final int maxValue = 10;
 
         int randomValue = ThreadLocalRandom.current().nextInt(minValue, maxValue + 1);
-        return quatidadeItensSelecionados * randomValue;
+        return (double) (quatidadeItensSelecionados * randomValue);
     }
 
 }
